(related to bug 886) Correct the daemons no to be killed by exiting the shell... This is needed for ssh and daemon start with user elab (mainly for tests). Also, try to detect if we are root user or the desired user allready to avoid issueing SU command. Correct the calculations of the delta times in microsseconds, as we didn't take into account the differences in seconds and minutes and even more. Provide a framework level method for reuse. Change the scales of the output channels according to the new calculations of wave form values and also of the intensity average power level. Change the calculation method of the wave form sample values and the normalization factor to be dependent on the audioFormat. Debug more information about the received Buffers to check if there is too much delay to the received data as it looks currently by running the experiment.
git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@5093 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ rec/src/java/common/com/linkare/rec/jmf/media/datasink/capture/ChannelDataFrame.java
@@ -4,21 +4,21 @@ import javax.media.format.AudioFormat;
 
 public class ChannelDataFrame {
 
-	private long[][] channelsData;
+	private double[][] channelsData;
 	private double[] channelsVRMS;
 	private AudioFormat captureFormat;
 
-	public ChannelDataFrame(long[][] channelsData, double[] channelsVRMS, AudioFormat captureFormat) {
+	public ChannelDataFrame(double[][] channelsData, double[] channelsVRMS, AudioFormat captureFormat) {
 		this.channelsData = channelsData;
 		this.channelsVRMS = channelsVRMS;
 		this.captureFormat = captureFormat;
 	}
 
-	public long[][] getChannelsData() {
+	public double[][] getChannelsData() {
 		return channelsData;
 	}
 
-	public void setChannelsData(long[][] channelsData) {
+	public void setChannelsData(double[][] channelsData) {
 		this.channelsData = channelsData;
 	}
 
@@ -34,7 +34,7 @@ public class ChannelDataFrame {
 		return channelsData.length;
 	}
 
-	public long[] getChannelData(int channel) {
+	public double[] getChannelData(int channel) {
 		return channelsData[channel];
 	}
